services/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
sqlalchemy
aiosqlite
PyJWT
pytest
